cleanup go code
