+++ git-https-proxy/config_old/config_test.go
